moved stop_managed_acquisition route and others over to new server
