+++ tests/test_main_route_queue.py
@@ -7,7 +7,6 @@ from mantarray_desktop_app import ADC_GAIN_SETTING_UUID
 from mantarray_desktop_app import COMPILED_EXE_BUILD_TIMESTAMP
 from mantarray_desktop_app import CURRENT_SOFTWARE_VERSION
 from mantarray_desktop_app import CUSTOMER_ACCOUNT_ID_UUID
-from mantarray_desktop_app import main
 from mantarray_desktop_app import MAIN_FIRMWARE_VERSION_UUID
 from mantarray_desktop_app import MANTARRAY_NICKNAME_UUID
 from mantarray_desktop_app import MANTARRAY_SERIAL_NUMBER_UUID
@@ -257,137 +256,6 @@ def test_start_recording_command__populates_queue__with_correct_adc_offset_value
     )
 
 
-def test_send_single_read_wire_out_command__populates_queue__and_logs_response(
-    test_process_manager, test_client, mocker
-):
-    board_idx = 0
-    expected_ep_addr = 6
-    mocked_logger = mocker.patch.object(main.logger, "info", autospec=True)
-    response = test_client.get(
-        f"/insert_xem_command_into_queue/read_wire_out?ep_addr={expected_ep_addr}"
-    )
-    assert response.status_code == 200
-
-    comm_queue = test_process_manager.get_communication_to_ok_comm_queue(board_idx)
-    assert is_queue_eventually_not_empty(comm_queue) is True
-    communication = comm_queue.get_nowait()
-    assert communication["communication_type"] == "debug_console"
-    assert communication["command"] == "read_wire_out"
-    assert communication["ep_addr"] == expected_ep_addr
-    assert communication["suppress_error"] is True
-    response_json = response.get_json()
-    assert response_json["command"] == "read_wire_out"
-    assert response_json["ep_addr"] == expected_ep_addr
-    assert response_json["suppress_error"] is True
-    mocked_logger.assert_called_once_with(
-        f"Response to HTTP Request in next log entry: {response.get_json()}"
-    )
-
-
-def test_send_single_read_wire_out_command_with_hex_notation__populates_queue(
-    test_process_manager, test_client
-):
-    board_idx = 0
-    expected_ep_addr = "0x6"
-    response = test_client.get(
-        f"/insert_xem_command_into_queue/read_wire_out?ep_addr={expected_ep_addr}"
-    )
-    assert response.status_code == 200
-
-    comm_queue = test_process_manager.get_communication_to_ok_comm_queue(board_idx)
-    assert is_queue_eventually_not_empty(comm_queue) is True
-    communication = comm_queue.get_nowait()
-    assert communication["communication_type"] == "debug_console"
-    assert communication["command"] == "read_wire_out"
-    assert communication["ep_addr"] == 6
-    assert communication["suppress_error"] is True
-    response_json = response.get_json()
-    assert response_json["command"] == "read_wire_out"
-    assert response_json["ep_addr"] == 6
-    assert response_json["suppress_error"] is True
-
-
-def test_send_single_read_wire_out_command_with_description__populates_queue(
-    test_process_manager, test_client
-):
-    board_idx = 0
-    expected_ep_addr = 6
-    expected_description = "test"
-    response = test_client.get(
-        f"/insert_xem_command_into_queue/read_wire_out?ep_addr={expected_ep_addr}&description={expected_description}"
-    )
-    assert response.status_code == 200
-
-    comm_queue = test_process_manager.get_communication_to_ok_comm_queue(board_idx)
-    assert is_queue_eventually_not_empty(comm_queue) is True
-    communication = comm_queue.get_nowait()
-    assert communication["communication_type"] == "debug_console"
-    assert communication["command"] == "read_wire_out"
-    assert communication["ep_addr"] == expected_ep_addr
-    assert communication["description"] == expected_description
-    assert communication["suppress_error"] is True
-    response_json = response.get_json()
-    assert response_json["command"] == "read_wire_out"
-    assert response_json["ep_addr"] == expected_ep_addr
-    assert response_json["description"] == expected_description
-    assert response_json["suppress_error"] is True
-
-
-def test_send_single_set_wire_in_command__populates_queue(
-    test_process_manager, test_client
-):
-    expected_ep_addr = 8
-    expected_value = 0x00000010
-    expected_mask = 0x00000010
-    response = test_client.get(
-        f"/insert_xem_command_into_queue/set_wire_in?ep_addr={expected_ep_addr}&value={expected_value}&mask={expected_mask}"
-    )
-    assert response.status_code == 200
-    comm_queue = test_process_manager.get_communication_to_ok_comm_queue(0)
-    assert is_queue_eventually_not_empty(comm_queue) is True
-    communication = comm_queue.get_nowait()
-    assert communication["communication_type"] == "debug_console"
-    assert communication["command"] == "set_wire_in"
-    assert communication["ep_addr"] == expected_ep_addr
-    assert communication["value"] == expected_value
-    assert communication["mask"] == expected_mask
-    assert communication["suppress_error"] is True
-    response_json = response.get_json()
-    assert response_json["command"] == "set_wire_in"
-    assert response_json["ep_addr"] == expected_ep_addr
-    assert response_json["value"] == expected_value
-    assert response_json["mask"] == expected_mask
-    assert response_json["suppress_error"] is True
-
-
-def test_send_single_set_wire_in_command__using_hex_notation__populates_queue(
-    test_process_manager, test_client
-):
-    expected_ep_addr = "0x05"
-    value = "0x000000a0"
-    mask = "0x00000011"
-    response = test_client.get(
-        f"/insert_xem_command_into_queue/set_wire_in?ep_addr={expected_ep_addr}&value={value}&mask={mask}"
-    )
-    assert response.status_code == 200
-
-    comm_queue = test_process_manager.get_communication_to_ok_comm_queue(0)
-    assert is_queue_eventually_not_empty(comm_queue) is True
-    communication = comm_queue.get_nowait()
-    assert communication["communication_type"] == "debug_console"
-    assert communication["command"] == "set_wire_in"
-    assert communication["ep_addr"] == 5
-    assert communication["value"] == 160
-    assert communication["mask"] == 17
-    assert communication["suppress_error"] is True
-    response_json = response.get_json()
-    assert response_json["command"] == "set_wire_in"
-    assert response_json["ep_addr"] == 5
-    assert response_json["value"] == 160
-    assert response_json["mask"] == 17
-    assert response_json["suppress_error"] is True
-
-
 def test_send_single_start_managed_acquisition_command__populates_queues(
     test_process_manager, test_client, patched_shared_values_dict
 ):
@@ -414,57 +282,6 @@ def test_send_single_start_managed_acquisition_command__populates_queues(
     assert comm_to_da["command"] == "start_managed_acquisition"
     response_json = response.get_json()
     assert response_json["command"] == "start_managed_acquisition"
-
-
-def test_send_single_stop_managed_acquisition_command__populates_queues(
-    test_process_manager, test_client
-):
-    response = test_client.get("/stop_managed_acquisition")
-    assert response.status_code == 200
-
-    to_ok_comm_queue = test_process_manager.get_communication_to_ok_comm_queue(0)
-    assert is_queue_eventually_not_empty(to_ok_comm_queue) is True
-    comm_to_ok_comm = to_ok_comm_queue.get_nowait()
-    assert comm_to_ok_comm["communication_type"] == "acquisition_manager"
-    assert comm_to_ok_comm["command"] == "stop_managed_acquisition"
-    response_json = response.get_json()
-    assert response_json["command"] == "stop_managed_acquisition"
-
-    to_file_writer_queue = (
-        test_process_manager.get_communication_queue_from_main_to_file_writer()
-    )
-    assert is_queue_eventually_not_empty(to_file_writer_queue) is True
-    comm_to_da = to_file_writer_queue.get_nowait()
-    assert comm_to_da["communication_type"] == "acquisition_manager"
-    assert comm_to_da["command"] == "stop_managed_acquisition"
-    response_json = response.get_json()
-    assert response_json["command"] == "stop_managed_acquisition"
-
-    to_da_queue = (
-        test_process_manager.get_communication_queue_from_main_to_data_analyzer()
-    )
-    assert is_queue_eventually_not_empty(to_da_queue) is True
-    comm_to_da = to_da_queue.get_nowait()
-    assert comm_to_da["communication_type"] == "acquisition_manager"
-    assert comm_to_da["command"] == "stop_managed_acquisition"
-    response_json = response.get_json()
-    assert response_json["command"] == "stop_managed_acquisition"
-
-
-def test_send_single_xem_scripts_command__populates_queue(
-    test_process_manager, test_client
-):
-    expected_script_type = "start_up"
-    response = test_client.get(f"/xem_scripts?script_type={expected_script_type}")
-    assert response.status_code == 200
-
-    comm_queue = test_process_manager.get_communication_to_ok_comm_queue(0)
-    assert is_queue_eventually_not_empty(comm_queue) is True
-    communication = comm_queue.get_nowait()
-    assert communication["communication_type"] == "xem_scripts"
-    assert communication["script_type"] == expected_script_type
-    response_json = response.get_json()
-    assert response_json["script_type"] == expected_script_type
 
 
 def test_send_single_boot_up_command__populates_queue_with_both_commands(
