+++ tests/test_server/test_server_routes_populate_queues.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+from mantarray_desktop_app import server
+
 from ..fixtures import fixture_generic_queue_container
 from ..fixtures import QUEUE_CHECK_TIMEOUT_SECONDS
 from ..fixtures_server import fixture_client_and_server_thread_and_shared_values
@@ -345,16 +347,6 @@ def test_send_single_read_from_fifo_command__populates_queue(
     test_client, test_server_info, _ = client_and_server_thread_and_shared_values
     test_server, _, _ = test_server_info
 
-    # test_bytearray = produce_data(1, 0)
-    # fifo = Queue()
-    # fifo.put(test_bytearray)
-    # queues = {"pipe_outs": {PIPE_OUT_FIFO: fifo}}
-    # simulator = FrontPanelSimulator(queues)
-    # simulator.initialize_board()
-    # simulator.start_acquisition()
-    # ok_process = test_process_manager.get_ok_comm_process()
-    # ok_process.set_board_connection(0, simulator)
-
     response = test_client.get(
         "/insert_xem_command_into_queue/read_from_fifo?num_words_to_log=72"
     )
@@ -376,16 +368,6 @@ def test_send_single_read_from_fifo_command_with_hex_notation__populates_queue(
     test_client, test_server_info, _ = client_and_server_thread_and_shared_values
     test_server, _, _ = test_server_info
 
-    # test_bytearray = produce_data(1, 0)
-    # fifo = Queue()
-    # fifo.put(test_bytearray)
-    # queues = {"pipe_outs": {PIPE_OUT_FIFO: fifo}}
-    # simulator = FrontPanelSimulator(queues)
-    # simulator.initialize_board()
-    # simulator.start_acquisition()
-    # ok_process = test_process_manager.get_ok_comm_process()
-    # ok_process.set_board_connection(0, simulator)
-
     response = test_client.get(
         "/insert_xem_command_into_queue/read_from_fifo?num_words_to_log=0x48"
     )
@@ -400,3 +382,214 @@ def test_send_single_read_from_fifo_command_with_hex_notation__populates_queue(
     response_json = response.get_json()
     assert response_json["command"] == "read_from_fifo"
     assert response_json["suppress_error"] is True
+
+
+def test_send_single_set_wire_in_command__populates_queue(
+    client_and_server_thread_and_shared_values,
+):
+    test_client, test_server_info, _ = client_and_server_thread_and_shared_values
+    test_server, _, _ = test_server_info
+
+    expected_ep_addr = 8
+    expected_value = 0x00000010
+    expected_mask = 0x00000010
+    response = test_client.get(
+        f"/insert_xem_command_into_queue/set_wire_in?ep_addr={expected_ep_addr}&value={expected_value}&mask={expected_mask}"
+    )
+    assert response.status_code == 200
+    comm_queue = test_server.queue_container().get_communication_to_ok_comm_queue(0)
+    assert is_queue_eventually_not_empty(comm_queue) is True
+    communication = comm_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert communication["communication_type"] == "debug_console"
+    assert communication["command"] == "set_wire_in"
+    assert communication["ep_addr"] == expected_ep_addr
+    assert communication["value"] == expected_value
+    assert communication["mask"] == expected_mask
+    assert communication["suppress_error"] is True
+    response_json = response.get_json()
+    assert response_json["command"] == "set_wire_in"
+    assert response_json["ep_addr"] == expected_ep_addr
+    assert response_json["value"] == expected_value
+    assert response_json["mask"] == expected_mask
+    assert response_json["suppress_error"] is True
+
+
+def test_send_single_set_wire_in_command__using_hex_notation__populates_queue(
+    client_and_server_thread_and_shared_values,
+):
+    test_client, test_server_info, _ = client_and_server_thread_and_shared_values
+    test_server, _, _ = test_server_info
+
+    expected_ep_addr = "0x05"
+    value = "0x000000a0"
+    mask = "0x00000011"
+    response = test_client.get(
+        f"/insert_xem_command_into_queue/set_wire_in?ep_addr={expected_ep_addr}&value={value}&mask={mask}"
+    )
+    assert response.status_code == 200
+
+    comm_queue = test_server.queue_container().get_communication_to_ok_comm_queue(0)
+    assert is_queue_eventually_not_empty(comm_queue) is True
+    communication = comm_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert communication["communication_type"] == "debug_console"
+    assert communication["command"] == "set_wire_in"
+    assert communication["ep_addr"] == 5
+    assert communication["value"] == 160
+    assert communication["mask"] == 17
+    assert communication["suppress_error"] is True
+    response_json = response.get_json()
+    assert response_json["command"] == "set_wire_in"
+    assert response_json["ep_addr"] == 5
+    assert response_json["value"] == 160
+    assert response_json["mask"] == 17
+    assert response_json["suppress_error"] is True
+
+
+def test_send_single_xem_scripts_command__populates_queue(
+    client_and_server_thread_and_shared_values,
+):
+    test_client, test_server_info, _ = client_and_server_thread_and_shared_values
+    test_server, _, _ = test_server_info
+
+    expected_script_type = "start_up"
+    response = test_client.get(f"/xem_scripts?script_type={expected_script_type}")
+    assert response.status_code == 200
+
+    comm_queue = test_server.queue_container().get_communication_to_ok_comm_queue(0)
+    assert is_queue_eventually_not_empty(comm_queue) is True
+    communication = comm_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert communication["communication_type"] == "xem_scripts"
+    assert communication["script_type"] == expected_script_type
+    response_json = response.get_json()
+    assert response_json["script_type"] == expected_script_type
+
+
+def test_send_single_read_wire_out_command__populates_queue__and_logs_response(
+    client_and_server_thread_and_shared_values, mocker
+):
+    test_client, test_server_info, _ = client_and_server_thread_and_shared_values
+    test_server, _, _ = test_server_info
+
+    board_idx = 0
+    expected_ep_addr = 6
+    mocked_logger = mocker.patch.object(server.logger, "info", autospec=True)
+    response = test_client.get(
+        f"/insert_xem_command_into_queue/read_wire_out?ep_addr={expected_ep_addr}"
+    )
+    assert response.status_code == 200
+
+    comm_queue = test_server.queue_container().get_communication_to_ok_comm_queue(
+        board_idx
+    )
+    assert is_queue_eventually_not_empty(comm_queue) is True
+    communication = comm_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert communication["communication_type"] == "debug_console"
+    assert communication["command"] == "read_wire_out"
+    assert communication["ep_addr"] == expected_ep_addr
+    assert communication["suppress_error"] is True
+    response_json = response.get_json()
+    assert response_json["command"] == "read_wire_out"
+    assert response_json["ep_addr"] == expected_ep_addr
+    assert response_json["suppress_error"] is True
+    mocked_logger.assert_called_once_with(
+        f"Response to HTTP Request in next log entry: {response.get_json()}"
+    )
+
+
+def test_send_single_read_wire_out_command_with_hex_notation__populates_queue(
+    client_and_server_thread_and_shared_values,
+):
+    test_client, test_server_info, _ = client_and_server_thread_and_shared_values
+    test_server, _, _ = test_server_info
+
+    board_idx = 0
+    expected_ep_addr = "0x6"
+    response = test_client.get(
+        f"/insert_xem_command_into_queue/read_wire_out?ep_addr={expected_ep_addr}"
+    )
+    assert response.status_code == 200
+
+    comm_queue = test_server.queue_container().get_communication_to_ok_comm_queue(
+        board_idx
+    )
+    assert is_queue_eventually_not_empty(comm_queue) is True
+    communication = comm_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert communication["communication_type"] == "debug_console"
+    assert communication["command"] == "read_wire_out"
+    assert communication["ep_addr"] == 6
+    assert communication["suppress_error"] is True
+    response_json = response.get_json()
+    assert response_json["command"] == "read_wire_out"
+    assert response_json["ep_addr"] == 6
+    assert response_json["suppress_error"] is True
+
+
+def test_send_single_read_wire_out_command_with_description__populates_queue(
+    client_and_server_thread_and_shared_values,
+):
+    test_client, test_server_info, _ = client_and_server_thread_and_shared_values
+    test_server, _, _ = test_server_info
+
+    board_idx = 0
+    expected_ep_addr = 6
+    expected_description = "test"
+    response = test_client.get(
+        f"/insert_xem_command_into_queue/read_wire_out?ep_addr={expected_ep_addr}&description={expected_description}"
+    )
+    assert response.status_code == 200
+
+    comm_queue = test_server.queue_container().get_communication_to_ok_comm_queue(
+        board_idx
+    )
+    assert is_queue_eventually_not_empty(comm_queue) is True
+    communication = comm_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert communication["communication_type"] == "debug_console"
+    assert communication["command"] == "read_wire_out"
+    assert communication["ep_addr"] == expected_ep_addr
+    assert communication["description"] == expected_description
+    assert communication["suppress_error"] is True
+    response_json = response.get_json()
+    assert response_json["command"] == "read_wire_out"
+    assert response_json["ep_addr"] == expected_ep_addr
+    assert response_json["description"] == expected_description
+    assert response_json["suppress_error"] is True
+
+
+def test_send_single_stop_managed_acquisition_command__populates_queues(
+    client_and_server_thread_and_shared_values,
+):
+    test_client, test_server_info, _ = client_and_server_thread_and_shared_values
+    test_server, _, _ = test_server_info
+
+    response = test_client.get("/stop_managed_acquisition")
+    assert response.status_code == 200
+
+    to_ok_comm_queue = test_server.queue_container().get_communication_to_ok_comm_queue(
+        0
+    )
+    assert is_queue_eventually_not_empty(to_ok_comm_queue) is True
+    comm_to_ok_comm = to_ok_comm_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert comm_to_ok_comm["communication_type"] == "acquisition_manager"
+    assert comm_to_ok_comm["command"] == "stop_managed_acquisition"
+    response_json = response.get_json()
+    assert response_json["command"] == "stop_managed_acquisition"
+
+    to_file_writer_queue = (
+        test_server.queue_container().get_communication_queue_from_main_to_file_writer()
+    )
+    assert is_queue_eventually_not_empty(to_file_writer_queue) is True
+    comm_to_da = to_file_writer_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert comm_to_da["communication_type"] == "acquisition_manager"
+    assert comm_to_da["command"] == "stop_managed_acquisition"
+    response_json = response.get_json()
+    assert response_json["command"] == "stop_managed_acquisition"
+
+    to_da_queue = (
+        test_server.queue_container().get_communication_queue_from_main_to_data_analyzer()
+    )
+    assert is_queue_eventually_not_empty(to_da_queue) is True
+    comm_to_da = to_da_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert comm_to_da["communication_type"] == "acquisition_manager"
+    assert comm_to_da["command"] == "stop_managed_acquisition"
+    response_json = response.get_json()
+    assert response_json["command"] == "stop_managed_acquisition"
